move 2d tests
